+++ platform-operator/experimental/controllers/verrazzano/custom/rancher.go
@@ -1,14 +1,16 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package verrazzano
+package custom
 
 import (
 	"context"
-	"strings"
-
+	"github.com/verrazzano/verrazzano-modules/pkg/controller/result"
 	"github.com/verrazzano/verrazzano/pkg/constants"
+	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	vzv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	vzconst "github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/rancher"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
@@ -18,14 +20,16 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"strings"
 )
 
-// createRancherIngressAndCertCopies - creates copies of Rancher ingress and cert secret if
+// CreateRancherIngressAndCertCopies - creates copies of Rancher ingress and cert secret if
 // they exist. If there is a DNS update in progress, we want to do this so that the old DNS
 // continues to work as long as there is a good DNS entry, so that managed clusters can continue
 // connecting using the old Rancher DNS in their kubeconfig, until they are updated with the new DNS.
-func (r *Reconciler) createRancherIngressAndCertCopies(ctx spi.ComponentContext) {
+func CreateRancherIngressAndCertCopies(ctx spi.ComponentContext) {
 	ok, rancherComp := registry.FindComponent(rancher.ComponentName)
 	if !ok {
 		return
@@ -39,7 +43,7 @@ func (r *Reconciler) createRancherIngressAndCertCopies(ctx spi.ComponentContext)
 
 	for _, ingressName := range rancherComp.GetIngressNames(ctx) {
 		ing := netv1.Ingress{}
-		err := r.Client.Get(context.TODO(), ingressName, &ing)
+		err := ctx.Client().Get(context.TODO(), ingressName, &ing)
 		if err != nil {
 			if !errors.IsNotFound(err) {
 				ctx.Log().Infof("Cannot make a copy of Rancher ingress %s/%s - get ingress failed: %v", ingressName.Namespace, ingressName.Name, err)
@@ -55,9 +59,9 @@ func (r *Reconciler) createRancherIngressAndCertCopies(ctx spi.ComponentContext)
 			continue
 		}
 
-		newIngressTLSList := r.createIngressCertSecretCopies(ctx, ing)
+		newIngressTLSList := createIngressCertSecretCopies(ctx, ing)
 		newIngressNSN := types.NamespacedName{Name: "vz-" + ing.Name, Namespace: ingressName.Namespace}
-		err = r.createIngressCopy(newIngressNSN, ing, newIngressTLSList)
+		err = createIngressCopy(ctx.Client(), newIngressNSN, ing, newIngressTLSList)
 		if err != nil {
 			ctx.Log().Infof("Failed to create a copy of Rancher ingress %s/%s - create ingress failed: %v", ingressName.Namespace, ingressName.Name, err)
 			continue
@@ -66,14 +70,14 @@ func (r *Reconciler) createRancherIngressAndCertCopies(ctx spi.ComponentContext)
 	}
 }
 
-func (r *Reconciler) createIngressCopy(newIngressName types.NamespacedName, existingIngress netv1.Ingress, newIngressTLSList []netv1.IngressTLS) error {
+func createIngressCopy(cli client.Client, newIngressName types.NamespacedName, existingIngress netv1.Ingress, newIngressTLSList []netv1.IngressTLS) error {
 	ingress := &netv1.Ingress{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      newIngressName.Name,
 			Namespace: newIngressName.Namespace,
 		},
 	}
-	_, err := controllerutil.CreateOrUpdate(context.TODO(), r.Client, ingress, func() error {
+	_, err := controllerutil.CreateOrUpdate(context.TODO(), cli, ingress, func() error {
 		ingress.Labels = existingIngress.Labels
 		if ingress.Labels == nil {
 			ingress.Labels = map[string]string{}
@@ -87,16 +91,16 @@ func (r *Reconciler) createIngressCopy(newIngressName types.NamespacedName, exis
 	return err
 }
 
-func (r *Reconciler) createIngressCertSecretCopies(ctx spi.ComponentContext, ing netv1.Ingress) []netv1.IngressTLS {
+func createIngressCertSecretCopies(ctx spi.ComponentContext, ing netv1.Ingress) []netv1.IngressTLS {
 	newIngressTLSList := []netv1.IngressTLS{}
 	for _, ingTLS := range ing.Spec.TLS {
 		if ingTLS.SecretName != "" {
 			tlsSecret := corev1.Secret{}
 			tlsSecretName := types.NamespacedName{Namespace: ing.Namespace, Name: ingTLS.SecretName}
-			err := r.Client.Get(context.TODO(), tlsSecretName, &tlsSecret)
+			err := ctx.Client().Get(context.TODO(), tlsSecretName, &tlsSecret)
 			if err == nil {
 				newSecretNSN := types.NamespacedName{Name: "vz-" + tlsSecret.Name, Namespace: tlsSecret.Namespace}
-				if err := r.createSecretCopy(newSecretNSN, tlsSecret); err != nil {
+				if err := createSecretCopy(ctx.Client(), newSecretNSN, tlsSecret); err != nil {
 					ctx.Log().Infof("Failed to create copy %v of Rancher TLS secret %v", newSecretNSN, tlsSecretName)
 				}
 				newIngressTLS := ingTLS.DeepCopy()
@@ -108,14 +112,14 @@ func (r *Reconciler) createIngressCertSecretCopies(ctx spi.ComponentContext, ing
 	return newIngressTLSList
 }
 
-func (r *Reconciler) createSecretCopy(newName types.NamespacedName, existingSecret corev1.Secret) error {
+func createSecretCopy(cli client.Client, newName types.NamespacedName, existingSecret corev1.Secret) error {
 	newSecret := &corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: newName.Namespace,
 			Name:      newName.Name,
 		},
 	}
-	_, err := controllerutil.CreateOrUpdate(context.TODO(), r.Client, newSecret, func() error {
+	_, err := controllerutil.CreateOrUpdate(context.TODO(), cli, newSecret, func() error {
 		newSecret.Labels = existingSecret.Labels
 		newSecret.Annotations = existingSecret.Annotations
 		newSecret.Data = existingSecret.Data
@@ -136,4 +140,55 @@ func getDNSSuffix(effectiveCR *v1alpha1.Verrazzano) string {
 	}
 
 	return dnsSuffix
+}
+
+// preUninstallRancherLocal does Rancher pre-uninstall
+func PreUninstallRancher(cli client.Client, log vzlog.VerrazzanoLogger, actualCR *vzv1alpha1.Verrazzano, effectiveCR *vzv1alpha1.Verrazzano) result.Result {
+	rancherProvisioned, err := rancher.IsClusterProvisionedByRancher()
+	if err != nil {
+		return result.NewResultShortRequeueDelayWithError(err)
+	}
+
+	// Don't remove Rancher local namespace if cluster was provisioned by Rancher (for example RKE2).  Removing
+	// will cause cluster corruption.
+	if rancherProvisioned {
+		return result.NewResult()
+	}
+	// If Rancher is installed, then delete local cluster
+	found, comp := registry.FindComponent(rancher.ComponentName)
+	if !found {
+		return result.NewResult()
+	}
+
+	spiCtx, err := spi.NewContext(log, cli, actualCR, nil, false)
+	if err != nil {
+		return result.NewResultShortRequeueDelayWithError(err)
+	}
+	compContext := spiCtx.Init(rancher.ComponentName).Operation(vzconst.UninstallOperation)
+	installed, err := comp.IsInstalled(compContext)
+	if err != nil {
+		return result.NewResultShortRequeueDelayWithError(err)
+	}
+	if !installed {
+		return result.NewResult()
+	}
+	rancher.DeleteLocalCluster(log, cli)
+
+	if err := DeleteMCResources(spiCtx); err != nil {
+		return result.NewResultShortRequeueDelayWithError(err)
+	}
+	return result.NewResult()
+}
+
+func RunRancherPostUninstall(ctx spi.ComponentContext) error {
+	// Look up the Rancher component and call PostUninstall explicitly, without checking if it's installed;
+	// this is to catch any lingering managed cluster resources
+	if found, comp := registry.FindComponent(rancher.ComponentName); found {
+		err := comp.PostUninstall(ctx.Init(rancher.ComponentName).Operation(vzconst.UninstallOperation))
+		if err != nil {
+			ctx.Log().Once("Waiting for Rancher post-uninstall cleanup to be done")
+			return err
+		}
+	}
+	return nil
 }
