+++ platform-operator/experimental/controllers/verrazzano/custom/secret.go
@@ -1,13 +1,14 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package verrazzano
+package custom
 
 import (
 	"context"
 	"fmt"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	constants2 "github.com/verrazzano/verrazzano/pkg/mcconstants"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
@@ -22,8 +23,7 @@ import (
 // so that the admin cluster can manage multi-cluster resources without the need of a VMC.  In the
 // case of a cluster that is used as a managed cluster, this secret will still be created, but not used, and
 // ultimately replaced when the user applies the managed cluster YAML which has the replacement secret.
-func (r *Reconciler) syncLocalRegistrationSecret() error {
-
+func SyncLocalRegistrationSecret(cli client.Client) error {
 	// If the agent secret exists in verrazzano-system namespace then that means this
 	// is a managed cluster and the user applied the YAML to create all the secrets.
 	// In that case, we do NOT want to create/update this default local secret since
@@ -34,7 +34,7 @@ func (r *Reconciler) syncLocalRegistrationSecret() error {
 		Name:      constants.MCAgentSecret,
 	}
 	// get the agent secret and return if it exists
-	err := r.Client.Get(context.TODO(), nsn, &secret)
+	err := cli.Get(context.TODO(), nsn, &secret)
 	if err == nil {
 		return nil
 	}
@@ -43,7 +43,7 @@ func (r *Reconciler) syncLocalRegistrationSecret() error {
 	}
 
 	// create the local registration secret
-	_, err = r.createOrUpdateLocalRegistrationSecret(constants.MCLocalRegistrationSecret, constants.VerrazzanoSystemNamespace)
+	_, err = createOrUpdateLocalRegistrationSecret(cli, constants.MCLocalRegistrationSecret, constants.VerrazzanoSystemNamespace)
 	if err != nil {
 		return err
 	}
@@ -52,13 +52,13 @@ func (r *Reconciler) syncLocalRegistrationSecret() error {
 }
 
 // Create or update the secret
-func (r *Reconciler) createOrUpdateLocalRegistrationSecret(name string, namespace string) (controllerutil.OperationResult, error) {
+func createOrUpdateLocalRegistrationSecret(cli client.Client, name string, namespace string) (controllerutil.OperationResult, error) {
 	var secret corev1.Secret
 	secret.Namespace = namespace
 	secret.Name = name
 
-	return controllerutil.CreateOrUpdate(context.TODO(), r.Client, &secret, func() error {
-		r.mutateLocalRegistrationSecret(&secret)
+	return controllerutil.CreateOrUpdate(context.TODO(), cli, &secret, func() error {
+		mutateLocalRegistrationSecret(&secret)
 		// Verrrazzano resource cannot own this secret since it is in a different namespace
 		// The secret will get deleted when verrazzano-system namespace is deleted.
 		return nil
@@ -66,7 +66,7 @@ func (r *Reconciler) createOrUpdateLocalRegistrationSecret(name string, namespac
 }
 
 // Mutate the secret, setting the kubeconfig data
-func (r *Reconciler) mutateLocalRegistrationSecret(secret *corev1.Secret) error {
+func mutateLocalRegistrationSecret(secret *corev1.Secret) error {
 	secret.Type = corev1.SecretTypeOpaque
 	secret.Data = map[string][]byte{
 		constants2.ManagedClusterNameKey: []byte(constants.MCLocalCluster),
@@ -74,13 +74,13 @@ func (r *Reconciler) mutateLocalRegistrationSecret(secret *corev1.Secret) error 
 	return nil
 }
 
-// deleteSecret deletes a Kubernetes secret
-func (r *Reconciler) deleteSecret(log vzlog.VerrazzanoLogger, namespace string, name string) error {
+// DeleteSecret deletes a Kubernetes secret
+func DeleteSecret(cli client.Client, log vzlog.VerrazzanoLogger, namespace string, name string) error {
 	secret := corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name},
 	}
 	log.Oncef("Deleting multicluster secret %s:%s", namespace, name)
-	if err := r.Client.Delete(context.TODO(), &secret); err != nil {
+	if err := cli.Delete(context.TODO(), &secret); err != nil {
 		if errors.IsNotFound(err) {
 			return nil
 		}
